pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "pyairtable>=3.3.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]
